Change widget name to DSTicketMessage
