+++ lib/src/enums/ds_ticket_message_type.enum.dart
@@ -1,5 +1,5 @@
 /// A Design System enum that can be used to define the message type of the information
-enum DSInformationMessageType {
+enum DSTicketMessageType {
   forwardedTicket,
   closedCustomer,
   closedAttendant,
