+++ sample/lib/widgets/showcase/sample_ticket_message.showcase.dart
@@ -2,26 +2,26 @@ import 'package:flutter/material.dart';
 
 import 'package:blip_ds/blip_ds.dart';
 
-class SampleInformationMessage extends StatelessWidget {
-  const SampleInformationMessage({Key? key}) : super(key: key);
+class SampleTicketMessage extends StatelessWidget {
+  const SampleTicketMessage({Key? key}) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
     return Column(
       children: [
-        DSInformationMessage(
-          messageType: DSInformationMessageType.forwardedTicket,
+        DSTicketMessage(
+          messageType: DSTicketMessageType.forwardedTicket,
           ticketNumber: '#2022000123',
           chatbotIdentity: 'Papagaio',
         ),
         const SizedBox(height: 20.0),
-        DSInformationMessage(
-          messageType: DSInformationMessageType.closedAttendant,
+        DSTicketMessage(
+          messageType: DSTicketMessageType.closedAttendant,
           agentIdentity: 'Jhon Doe',
         ),
         const SizedBox(height: 20.0),
-        DSInformationMessage(
-          messageType: DSInformationMessageType.closedCustomer,
+        DSTicketMessage(
+          messageType: DSTicketMessageType.closedCustomer,
         ),
         const SizedBox(height: 20.0),
       ],
